Use int32 for the step counter instead of FP32
FP32 represents integers exactly only up to 2**24, so an FP32 counter has the
same failure mode as the BF16 one, just at 16,777,216 steps instead of 256.
int32 is exact to 2**31-1 and removes the saturation entirely rather than moving
it further out.

The consumption side is unchanged: beta**step returns the same FP32 value
bit-for-bit with an int32 step, and torch.compile(fullgraph=True) accepts it.
Checkpoints written before this change are unaffected -- a saved FP32 step loads
back as FP32 and keeps counting.

+++ torchao/optim/adam.py
@@ -121,11 +121,12 @@ class _AdamBase(Optimizer):
 
                     # State initialization
                     if len(state) == 0:
-                        # Pin the step counter to FP32 instead of letting it pick up the
-                        # default dtype. Under torch.set_default_dtype(torch.bfloat16) the
-                        # counter would be BF16, which cannot represent 257, so it saturates
-                        # at 256 and the bias corrections below freeze with it.
-                        state["step"] = torch.tensor(0.0, dtype=torch.float32)
+                        # Keep the step counter integral instead of letting it pick up the
+                        # default dtype. A float counter silently stops advancing once the
+                        # increment falls below one ULP -- at 256 under
+                        # torch.set_default_dtype(torch.bfloat16), and at 2**24 even in FP32 --
+                        # freezing the bias corrections derived from it below.
+                        state["step"] = torch.tensor(0, dtype=torch.int32)
                         state["exp_avg"] = self._new_buffer(p, True)
                         state["exp_avg_sq"] = self._new_buffer(p, False)
                         if group["amsgrad"]:
